instapic_api/requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.5.0
django-environ==0.4.5
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
